+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## Added
 
 - Accessibility changes added.
+- Seperated selected facets config
+- Github action
 
 
 ## [2.0.1] - 2020-12-22